LayoutTests/inspector/protocol/page-agent.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/protocol-test.js"></script>
<script>

function test()
{
    var testSuite = [
        ["PageAgent", "disable"],
        ["PageAgent", "enable"],
        ["PageAgent", "getResourceTree"],
    ];

    InspectorTest.runProtocolTestSuite("PageAgent", testSuite, ['id', 'parentId', 'loaderId', 'url']);
}

</script>
</head>

<body onload="runTest()">
<p>
Protocol stability test. It is dumping request/response pairs of PageAgent functions.
</p>

<img src="resources/empty.png">

<img src="resources/non-existent.png">

<style>
@font-face {
    font-family: test1;
    src: url(resources/font1.ttf);
}

@font-face {
    font-family: test2;
    src: url(resources/font2.ttf);
}

body {
    font-family: test2;
}
</style>

<iframe id="dummy" src="resources/dummy-iframe.html"></iframe>

</body>
</html>
